requirements.txt
open3d>=0.9.0.0
PyYAML==5.4.1
tqdm~=4.62.3
tensorboardX~=2.5.1
termcolor==1.1.0
scikit-image~=0.16.2
opencv-contrib-python>=3.4.2.17
opencv-python~=4.6.0.66
imageio~=2.22.1
trimesh~=3.15.3
plyfile==0.6
pyglet==1.4.0b1
chumpy
numpy~=1.23.0
matplotlib~=3.5.3
yacs~=0.1.6
Pillow~=9.2.0
PyMCubes==0.1.4
